Pushover notifications & bug fixes.
Fixed issue with  Pushover user key not being in base64.
Fixed NOTIFICATION being overwritten accidentally in noip-renew.sh, which caused no token parameters to be passed when notifications were switched on.
Finished work on Pushover Notifications - These are now the only notifications that will work properly. Guide in readme will be changed to reflect this.